myblog/requirements.txt
asgiref~=2.1
daphne==2.0.0
channels==2.0.0
beautifulsoup4==4.6.0
channels-redis==2.3.3
Django==2.2.0
django-allauth==0.41.0
django-appconf==1.0.3
django-ckeditor==5.9.0
django-imagekit==4.0.2
django-js-asset==1.2.2
django-mailgun==0.9.1
django-model-utils==4.0.0
django-mptt==0.11.0
django-notifications-hq==1.5.0
django-password-reset==2.0
django-redis==4.10.0
django-sslserver==0.22
django-taggit==1.2.0
docutils==0.14
dwebsocket==0.5.12
Markdown==3.1.1
numpy==1.17.3
pandas==0.25.1
pandocfilters==1.4.2
PyMySQL==0.9.3
dj-static==0.0.6
dj-database-url==0.5.0
dj-static==0.0.6
gunicorn==19.9.0
mysql-connector==2.2.9
mysql-connector-python==8.0.17
mysqlclient==1.4.6
Twisted==19.7.0
Braintree==4.0.0
pilkit==2.0
Pillow==5.0.0
psycopg2==2.8.4
zhconv==1.4.0
lxml==4.5.0